requirements.txt
swarmauri[full]==0.4.1
python-dotenv
gradio
groq
